Backport m4a.c from emerald
