+++ Makefile
@@ -59,8 +59,7 @@ LD_SCRIPT := $(BUILD_DIR)/ld_script.ld
 %src/libs/agb_flash.o:    CC1FLAGS := -O1 -mthumb-interwork
 %src/libs/agb_flash_1m.o: CC1FLAGS := -O1 -mthumb-interwork
 %src/libs/agb_flash_mx.o: CC1FLAGS := -O1 -mthumb-interwork
-%src/libs/m4a_2.o: CC1 := tools/agbcc/bin/old_agbcc$(EXE)
-%src/libs/m4a_4.o: CC1 := tools/agbcc/bin/old_agbcc$(EXE)
+%src/libs/m4a.o: CC1 := tools/agbcc/bin/old_agbcc$(EXE)
 %src/libs/libisagbprn.o: CC1 := tools/agbcc/bin/old_agbcc$(EXE)
 %src/libs/libisagbprn.o: CC1FLAGS := -mthumb-interwork
