+++ ld_script.txt
@@ -23,7 +23,7 @@ SECTIONS {
         INCLUDE "sym_bss.ld"
 
         /* .bss.code starts at 0x3000F60 */
-        src/libs/m4a_2.o(.bss.code);
+        src/libs/m4a.o(.bss.code);
 
         /* COMMON starts at 0x3001760 */
         INCLUDE "sym_common.ld"
@@ -400,9 +400,7 @@ SECTIONS {
     {
         src/libs/libgcnmultiboot.o(.text);
         src/libs/m4a_1.o(.text);
-        src/libs/m4a_2.o(.text);
-        asm/m4a_3.o(.text);
-        src/libs/m4a_4.o(.text);
+        src/libs/m4a.o(.text);
         src/libs/agb_flash.o(.text);
         src/libs/agb_flash_1m.o(.text);
         src/libs/agb_flash_mx.o(.text);
